requirements_backup.txt
Pillow
tqdm
h5py>=2.10.0
Flask
Keras>=2.3.1
Keras-Applications>=1.0.8
Keras-Preprocessing>=1.1.0
tensorflow==2.0.0
tensorflow-estimator>=2.0.1